+++ tests/numeric/test_LogTransformer.py
@@ -96,8 +96,8 @@ class TestTransform:
         """Expected output of test_expected_output_2."""
         df = d.create_df_3()
 
-        df["a_new_col"] = np.log(df["a"] + 1)
-        df["b_new_col"] = np.log(df["b"] + 1)
+        df["a_new_col"] = np.log1p(df["a"])
+        df["b_new_col"] = np.log1p(df["b"])
 
         return df.drop(columns=["a", "b"])
 
@@ -114,8 +114,8 @@ class TestTransform:
         """Expected output of test_expected_output_4."""
         df = d.create_df_3()
 
-        df["a_new_col"] = np.log(df["a"] + 1)
-        df["b_new_col"] = np.log(df["b"] + 1)
+        df["a_new_col"] = np.log1p(df["a"])
+        df["b_new_col"] = np.log1p(df["b"])
 
         return df
 
@@ -134,6 +134,28 @@ class TestTransform:
         df["a_new_col"] = np.log(df["a"]) / np.log(7)
 
         return df.drop("a", axis=1)
+
+    def test_log1p(self):
+        """Test that log1p is working as intended."""
+        df = pd.DataFrame(
+            {
+                "a": [0.00001, 0.00002, 0.00003],
+                "b": [0.00004, 0.00005, 0.00006],
+            },
+        )
+        # Values created using np.log1p() of original df
+        expected = pd.DataFrame(
+            {
+                "a_log": [9.999950e-06, 1.999980e-05, 2.999955e-05],
+                "b_log": [3.99992000e-05, 4.99987500e-05, 5.99982001e-05],
+            },
+        )
+        log_transformer = LogTransformer(
+            columns=["a", "b"],
+            add_1=True,
+        )
+        actual = log_transformer.transform(df)
+        pd.testing.assert_frame_equal(actual, expected)
 
     def test_super_transform_called(self, mocker):
         """Test that BaseTransformer.transform called."""