+++ CHANGELOG.rst
@@ -16,6 +16,17 @@ Subsections for each version can be one of the following;
 
 Each individual change should have a link to the pull request after the description of the change.
 
+1.2.2 (2024-02-20)
+------------------
+Added
+^^^^^
+- Created unit test for checking if log1p is working and well conditioned for small x `#178 <https://github.com/lvgig/tubular/pull/178>`_
+
+Changed
+^^^^^^^
+- Changed LogTransformer to use log1p(x) instead of log(x+1) `#178 <https://github.com/lvgig/tubular/pull/178>`_
+- Changed unit tests using log(x+1) to log1p(x) `#178 <https://github.com/lvgig/tubular/pull/178>`_
+
 1.2.1 (2024-02-08)
 ------------------
 Added